fix: take float score, better UX

+++ server/graph/model/models_gen.go
@@ -43,10 +43,10 @@ type Query struct {
 }
 
 type Rating struct {
-	ID    string `json:"id"`
-	User  *User  `json:"user"`
-	Movie *Movie `json:"movie"`
-	Score int    `json:"score"`
+	ID    string  `json:"id"`
+	User  *User   `json:"user"`
+	Movie *Movie  `json:"movie"`
+	Score float64 `json:"score"`
 }
 
 type User struct {
